Expose requestConfigCallback parameter in AwsElasticClient and set to public the visibility of the class AwsHttpClientConfigCallback (in this way it can be used directly with HttpClient)
