+++ elastic4s-aws/src/main/scala/com/sksamuel/elastic4s/aws/Aws4RequestSigner.scala
@@ -18,8 +18,8 @@ import org.apache.http.{Header, HttpRequest}
 class Aws4RequestSigner(provider: AWSCredentialsProvider, region: String, service: String = "es") {
   require(provider.getCredentials != null, "AWS Credentials are mandatory. AWSCredentialsProvider provided none.")
 
-  val dateHeader          = "X-Amz-Date"
-  val authHeader          = "Authorization"
+  val dateHeader = "X-Amz-Date"
+  val authHeader = "Authorization"
   val securityTokenHeader = "X-Amz-Security-Token"
 
   def withAws4Headers(request: HttpRequest): HttpRequest = {
@@ -34,7 +34,7 @@ class Aws4RequestSigner(provider: AWSCredentialsProvider, region: String, servic
     request.setHeader(dateHeader, dateTime)
 
     val canonicalRequest = CanonicalRequest(request)
-    val stringToSign     = StringToSign(service, region, canonicalRequest, date, dateTime)
+    val stringToSign = StringToSign(service, region, canonicalRequest, date, dateTime)
 
     val authHeaderValue = buildAuthenticationHeader(canonicalRequest, stringToSign, credentials)
     request.addHeader(authHeader, authHeaderValue)
@@ -42,7 +42,7 @@ class Aws4RequestSigner(provider: AWSCredentialsProvider, region: String, servic
     /* If the credentials are temporary (session credentials), add an additional security header */
     credentials match {
       case c: AWSSessionCredentials ⇒ request.addHeader(securityTokenHeader, c.getSessionToken)
-      case _                        ⇒
+      case _ ⇒
     }
 
     request
@@ -50,31 +50,31 @@ class Aws4RequestSigner(provider: AWSCredentialsProvider, region: String, servic
 
   /* Build date and dateTime in a protected method so it is possible to override it in tests */
   protected def buildDateAndDateTime(): (String, String) = {
-    val now      = ZonedDateTime.now(ZoneOffset.UTC)
+    val now = ZonedDateTime.now(ZoneOffset.UTC)
     val dateTime = now.format(DateTimeFormatter.ofPattern("yyyyMMdd'T'HHmmss'Z'"))
-    val date     = now.format(DateTimeFormatter.ofPattern("yyyyMMdd"))
+    val date = now.format(DateTimeFormatter.ofPattern("yyyyMMdd"))
     (date, dateTime)
   }
 
   private def buildAuthenticationHeader(canonicalRequest: CanonicalRequest,
                                         stringToSign: StringToSign,
                                         credentials: AWSCredentials) = {
-    val credentialStr    = s"Credential=${credentials.getAWSAccessKeyId}/${stringToSign.credentialsScope}"
+    val credentialStr = s"Credential=${credentials.getAWSAccessKeyId}/${stringToSign.credentialsScope}"
     val signedHeadersStr = s"SignedHeaders=${canonicalRequest.signedHeaders}"
-    val signatureStr     = s"Signature=${buildSignature(stringToSign, credentials)}"
+    val signatureStr = s"Signature=${buildSignature(stringToSign, credentials)}"
     s"${Crypto.Algorithm} $credentialStr, $signedHeadersStr, $signatureStr"
   }
 
   private def buildSignature(stringToSign: StringToSign, credentials: AWSCredentials) = {
     val signatureKey = buildKeyToSign(stringToSign.date, credentials)
-    val signature    = sign(stringToSign.toString, signatureKey)
+    val signature = sign(stringToSign.toString, signatureKey)
     hexOf(signature).toLowerCase
   }
 
   private def buildKeyToSign(dateStr: String, credentials: AWSCredentials): Array[Byte] = {
-    val kSecret    = ("AWS4" + credentials.getAWSSecretKey).getBytes("utf-8")
-    val dateKey    = sign(dateStr, kSecret)
-    val regionKey  = sign(region, dateKey)
+    val kSecret = ("AWS4" + credentials.getAWSSecretKey).getBytes("utf-8")
+    val dateKey = sign(dateStr, kSecret)
+    val regionKey = sign(region, dateKey)
     val serviceKey = sign(service, regionKey)
     sign("aws4_request", serviceKey)
   }
@@ -88,7 +88,7 @@ class Aws4RequestSigner(provider: AWSCredentialsProvider, region: String, servic
   private def setHeader(h: String, f: Header => String)(request: HttpRequest): HttpRequest = {
     request.getAllHeaders.find(_.getName == h) match {
       case Some(header) ⇒ request.setHeader(h, f(header))
-      case _            ⇒
+      case _ ⇒
     }
     request
   }
